MOVE: payorStatement code under rrpt
